+++ cvat-ui/src/actions/annotation-actions.ts
@@ -1133,7 +1133,7 @@ export function saveAnnotationsAsync(sessionInstance: any, afterSave?: () => voi
 
             if (sessionInstance instanceof cvat.classes.Job && sessionInstance.state === cvat.enums.JobState.NEW) {
                 sessionInstance.state = cvat.enums.JobState.IN_PROGRESS;
-                updateJobAsync(sessionInstance);
+                dispatch(updateJobAsync(sessionInstance));
             }
 
             dispatch({