+++ cvat/apps/iam/tests/test_rest_api.py
@@ -1,10 +1,10 @@
-# Copyright (C) 2021 Intel Corporation
+# Copyright (C) 2021-2022 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
 from django.urls import reverse
 from rest_framework import status
-from rest_framework.test import APITestCase
+from rest_framework.test import APITestCase, APIClient
 from rest_framework.authtoken.models import Token
 from django.test import override_settings
 from cvat.apps.iam.urls import urlpatterns as iam_url_patterns
@@ -19,14 +19,16 @@ urlpatterns = iam_url_patterns + [
          name='account_email_verification_sent'),
 ]
 
-
 class UserRegisterAPITestCase(APITestCase):
 
     user_data = {'first_name': 'test_first', 'last_name': 'test_last', 'username': 'test_username',
                  'email': 'test_email@test.com', 'password1': '$Test357Test%', 'password2': '$Test357Test%',
                  'confirmations': []}
 
-    def _run_api_v1_user_register(self, data):
+    def setUp(self):
+        self.client = APIClient()
+
+    def _run_api_v2_user_register(self, data):
         url = reverse('rest_register')
         response = self.client.post(url, data, format='json')
         return response
@@ -36,11 +38,11 @@ class UserRegisterAPITestCase(APITestCase):
         self.assertEqual(response.data, data)
 
     @override_settings(ACCOUNT_EMAIL_VERIFICATION='none')
-    def test_api_v1_user_register_with_email_verification_none(self):
+    def test_api_v2_user_register_with_email_verification_none(self):
         """
         Ensure we can register a user and get auth token key when email verification is none
         """
-        response = self._run_api_v1_user_register(self.user_data)
+        response = self._run_api_v2_user_register(self.user_data)
         user_token = Token.objects.get(user__username=response.data['username'])
         self._check_response(response, {'first_name': 'test_first', 'last_name': 'test_last',
                                         'username': 'test_username', 'email': 'test_email@test.com',
@@ -50,11 +52,11 @@ class UserRegisterAPITestCase(APITestCase):
     # the tests and pass it using ROOT_URLCONF in the override settings decorator
 
     @override_settings(ACCOUNT_EMAIL_VERIFICATION='optional', ROOT_URLCONF=__name__)
-    def test_api_v1_user_register_with_email_verification_optional(self):
+    def test_api_v2_user_register_with_email_verification_optional(self):
         """
         Ensure we can register a user and get auth token key when email verification is optional
         """
-        response = self._run_api_v1_user_register(self.user_data)
+        response = self._run_api_v2_user_register(self.user_data)
         user_token = Token.objects.get(user__username=response.data['username'])
         self._check_response(response, {'first_name': 'test_first', 'last_name': 'test_last',
                                         'username': 'test_username', 'email': 'test_email@test.com',
@@ -66,7 +68,7 @@ class UserRegisterAPITestCase(APITestCase):
         """
         Ensure we can register a user and it does not return auth token key when email verification is mandatory
         """
-        response = self._run_api_v1_user_register(self.user_data)
+        response = self._run_api_v2_user_register(self.user_data)
         self._check_response(response, {'first_name': 'test_first', 'last_name': 'test_last',
                                         'username': 'test_username', 'email': 'test_email@test.com',
                                         'email_verification_required': True, 'key': None})