+++ tests/cypress/integration/actions_projects_models/case_104_project_export_3d.js
@@ -1,4 +1,4 @@
-// Copyright (C) 2021 Intel Corporation
+// Copyright (C) 2021-2022 Intel Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -20,14 +20,12 @@ context('Export project dataset with 3D task.', { browser: '!firefox' }, () => {
         multiAttrParams: false,
     };
     let projectID = '';
+    let datasetArchiveName;
 
     function getProjectID() {
-        cy.contains('.cvat-project-name', projectName)
-            .parents('.cvat-project-details')
-            .should('have.attr', 'cvat-project-id')
-            .then(($projectID) => {
-                projectID = $projectID;
-            });
+        cy.url().then((url) => {
+            projectID = Number(url.split('/').slice(-1)[0].split('?')[0]);
+        });
     }
 
     before(() => {
@@ -74,7 +72,10 @@ context('Export project dataset with 3D task.', { browser: '!firefox' }, () => {
                 dumpType: 'Sly Point Cloud Format',
             };
             cy.exportProject(exportDataset3d);
-            cy.waitForDownload();
+            cy.getDownloadFileName().then((file) => {
+                datasetArchiveName = file;
+                cy.verifyDownload(datasetArchiveName);
+            });
         });
 
         it('Export project with 3D task. Annotation. Rename a archive.', () => {
@@ -83,11 +84,30 @@ context('Export project dataset with 3D task.', { browser: '!firefox' }, () => {
                 projectName,
                 as: 'exportAnnotations3dRenameArchive',
                 type: 'annotations',
-                dumpType: 'Kitti Raw Format',
+                dumpType: 'Datumaro 3D',
                 archiveCustomeName: 'export_project_3d_annotation',
             };
             cy.exportProject(exportAnnotations3dRenameArchive);
             cy.waitForDownload();
+        });
+
+        // FIXME: Activate after implementation
+        it.skip('Import dataset.', () => {
+            cy.openProject(projectName);
+            cy.deleteTask(task.name3d);
+            cy.get('.cvat-tasks-list-item')
+                .should('have.length', 1)
+                .should('have.attr', 'style')
+                .and('contain', 'pointer-events: none; opacity: 0.5;');
+            cy.goToProjectsList();
+            const importDataset = {
+                projectName,
+                format: 'Sly Point Cloud Format',
+                archive: datasetArchiveName,
+            };
+            cy.importProject(importDataset);
+            cy.openProject(projectName);
+            cy.get('.cvat-tasks-list-item').should('have.length', 1);
         });
     });
 });